requirements.txt
psutil
pyserial
pyinstaller